requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bitarray==2.9.3
bitstring==4.2.3
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
ecdsa==0.19.0
esptool==4.8.1
fastapi==0.115.12
fonttools==4.58.2
h11==0.16.0
idna==3.10
intelhex==2.3.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.2
reedsolo==1.7.0
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.5.0
tkinterdnd2==0.4.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
textual>=0.10.0
redis>=4.0.0
pytest>=7.0.0